My first try at runIn() to add a button held function
Because this is so different from what @tmleafs is working on, I've done as a Pull Request on his Pull.

I have no real way to test this on my Aqara button, so please have a look, and give it a try. I put in a bunch of debug log messages to help identify any issues that may arise.

I'm also looking into whether doing a sendEvent to change the state of "momentary" has any use. I notice it's in the old a4refillpad button DTH.