Add ignoreMotion countdown timer
As explained in my post on the SmartThings forum, I've added a second countdown timer so that all motion detected messages are ignored for a period of 5 seconds less than the motion reset timer. The state of ignore or not is tracked by a "global" state object, `state.ignoreMotion`.

This avoids issues with automations being activated every 5-6 seconds when the sensor is put into "testing mode."

I have also changed the default motion reset preference setting to 61 seconds, so that the motion reset counter can be extended another 60 seconds if the sensor detects movement right after its hardware "blind" period finishes.

I removed `state.battery = 0` from installed() and configure() because it's not used at all, and added `state.ignoreMotion == false` to those calls as well as `updated()` to ensure the DTH pairs the sensor or saves preference settings with motion detected messages NOT being ignored.